+++ src/main/java/subway/line/application/LineService.java
@@ -8,8 +8,6 @@ import subway.line.domain.Line;
 import subway.line.domain.LineCommandRepository;
 import subway.line.domain.LineQueryRepository;
 import subway.line.exception.LineNotFoundException;
-import subway.line.application.dto.request.SectionCreateRequest;
-import subway.line.domain.Section;
 import subway.station.application.StationService;
 import subway.station.domain.Station;
 
@@ -33,7 +31,7 @@ public class LineService {
 
     /**
      * 지하철 노선 목록 정보를 조회합니다.
-     * 
+     *
      * @return 지하철 노선 목록 정보
      */
     public List<Line> findAllLines() {
@@ -42,7 +40,7 @@ public class LineService {
 
     /**
      * 지하철 노선 상세 정보를 조횧합니다.
-     * 
+     *
      * @param lineId 지하철 노선 고유 번호
      * @return 지하철 노선 상세 정보
      */
@@ -68,7 +66,7 @@ public class LineService {
     /**
      * 지하철 노선 정보를 수정합니다.
      *
-     * @param lineId 수정할 지하철 노선 고유 번호
+     * @param lineId            수정할 지하철 노선 고유 번호
      * @param lineUpdateRequest 수정할 지하철 노선 정보
      */
     @Transactional
@@ -80,45 +78,12 @@ public class LineService {
 
     /**
      * 지하철 노선 정보를 삭제합니다.
-     * 
+     *
      * @param lineId 삭제할 지하철 노선 고유 번호
      */
     @Transactional
     public void deleteLine(final Long lineId) {
         lineCommandRepository.deleteById(lineId);
-    }
-
-    /**
-     * 지하철 구간 정보를 등록합니다.
-     * 
-     * @param lineId 등록할 지하철 구간의 노선 고유 번호
-     * @param sectionCreateRequest 등록할 지하철 구간 정보
-     * @return 등록된 지하철 구간 고유 번호
-     */
-    @Transactional
-    public Long saveSection(final Long lineId, final SectionCreateRequest sectionCreateRequest) {
-        Line findLine = findLineById(lineId);
-        Station upStation = stationService.findStationById(sectionCreateRequest.getUpStationId());
-        Station downStation = stationService.findStationById(sectionCreateRequest.getDownStationId());
-        Section section = Section.createSection(findLine, upStation, downStation, sectionCreateRequest.getDistance());
-
-        findLine.getSections().addSection(section);
-
-        return section.getId();
-    }
-
-    /**
-     * 지하철 구간 정보를 삭제합니다.
-     *
-     * @param lineId 삭제할 지하철 구간의 노선 고유 번호
-     * @param stationId 삭제할 지하철 구간 고유 번호
-     */
-    @Transactional
-    public void deleteSection(final Long lineId, final Long stationId) {
-        Line findLine = findLineById(lineId);
-        Station findStation = stationService.findStationById(stationId);
-
-        findLine.getSections().removeLastSection(findStation);
     }
 
     private Line createLine(final LineCreateRequest lineCreateRequest) {
