+++ src/main/java/subway/line/presentation/LineController.java
@@ -11,11 +11,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import subway.line.application.LineService;
+import subway.line.application.SectionService;
 import subway.line.application.dto.request.LineCreateRequest;
 import subway.line.application.dto.request.LineUpdateRequest;
+import subway.line.application.dto.request.SectionCreateRequest;
 import subway.line.application.dto.response.LineResponse;
 import subway.line.domain.Line;
-import subway.line.application.dto.request.SectionCreateRequest;
 
 import java.net.URI;
 import java.util.List;
@@ -25,9 +26,11 @@ import java.util.List;
 class LineController {
 
     private final LineService lineService;
+    private final SectionService sectionService;
 
-    LineController(final LineService lineService) {
+    LineController(final LineService lineService, final SectionService sectionService) {
         this.lineService = lineService;
+        this.sectionService = sectionService;
     }
 
     @PostMapping
@@ -73,7 +76,7 @@ class LineController {
     @PostMapping("/{lineId}/sections")
     ResponseEntity<Void> createSection(@PathVariable final Long lineId,
                                        @RequestBody final SectionCreateRequest sectionCreateRequest) {
-        lineService.saveSection(lineId, sectionCreateRequest);
+        sectionService.saveSection(lineId, sectionCreateRequest);
 
         return ResponseEntity.created(URI.create("/lines/" + lineId)).build();
     }
@@ -81,7 +84,7 @@ class LineController {
     @DeleteMapping("/{lineId}/sections")
     ResponseEntity<Void> deleteSection(@PathVariable final Long lineId,
                                        @RequestParam final Long stationId) {
-        lineService.deleteSection(lineId, stationId);
+        sectionService.deleteSection(lineId, stationId);
 
         return ResponseEntity.noContent().build();
     }
